requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
behave==1.2.6
blinker==1.9.0
cachelib==0.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
chess.com==3.11.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Limiter==3.10.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
ordered-set==4.1.0
packaging==24.2
parse==1.20.2
parse_type==0.6.4
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.38
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
gunicorn==21.2.0